Travis YAML: fix iojs version to 2.5.0
iojs 3.0.0 brings in multiple changes, one of which is the requirement
of more recent versions of build tools (gcc && friends). These, however,
are not available in Travis at the moment, causing build failures for
binary modules. This patch fixes the version to 2.5.0 to mitigate this
issue. In general, if/when we switch to iojs in WMF production, we'll be
using v2.5.0.

Also, removed testing on nodejs v0.11 as it's not used anywhere.

Change-Id: Ia8e7e22d8a0a4fef5f433a8a7cfe7f1c8bf24605