Add the 'complete' output format
This patch set adds the 'complete' output format which is equal to the
'json' one except that it also includes the headers for individual types
in the response body as well. The output specification is identical,
except that the 'mml', 'svg' and 'png' fields are now objects containing
the 'body' and 'headers' fields.

Note: this format is needed by RESTBase for
https://github.com/wikimedia/restbase/pull/339

Bug: T102030
Change-Id: I37d45cda1ceb5255a580b3a2f291268c1f270c53