Provide data to visualize the LaTeX AST
Expose the json data for the D3 formula visualisation data
via a new endpoint called graph. This data can thereafter
be visualized using graphoid. This will help editors to
understand how their formula is parsed.

In the future the same mechanism can be used as well to
visualize the formula tree.

Change-Id: I9c7432bee53694c34958cadff76c646580a31457