Disable svgo
mathoid has been tested intensively without svgo enabled.
T135955 shows that the SVG "optimization" is not loss-less.
Therefore, svgo is disabled for now.
To optionally serve compressed images, which is a requirement
for the the visual editor menu, we introduce the option svgo
(which is false by default) to generate svgo compressed images
on demand.
Note, that a developer has to start a mathoid instance on his
development machine with the option svg true to update
the visual editor menu using svgo compressed pictures.

Bug: T135955
Change-Id: I2abc30440de9ce262d0be88138855b60e24b8203