Add blubber.yaml
This is the file that will be used to build mathoid containers for
testing and for production.

This is a subset of Icccb0f79785e6112d4086da2d704d26e4144065e without
any configuration for helm. This creates a pipeline inside the dist
directory rather than putting a .blubber.yaml at the top level to allow
for future pipeline-related configuration.

Bug: T177954
Change-Id: I60b15718ec25dbb80fe2b205e610b632d55bdf35