Accept zlib compressed input
For the VMEXT research project, we can only send get requests to render
math. We have been using mathoids /get endpoint for that in the past.
This works fine in most cases. However, in some cases the URL is to
long.

The new endpoint /zlib accepts zlib compressed input and follows the
idea of T118028.

Change-Id: Ib7e4927cc9f4a6f882f2852a7b7f97361489d1e8