Amend swagger spec to conform to version 2.0
The version 2.0 swagger spec requires the responses are documented in
all endpoints, so add the 200 response to robots.txt, _info and root
endpoints. Add the 400 response as well to the POST /{format} endpoint
as per x-amples it's a valid response. Add a 200 response for the /zlib
endpoint

Document the "format" parameter of the /{format} endpoint. Use the enum
from outformat from the /zlib endpoint

Fix the /zlib endpoint path notation and set the obligatory required
attribute. Drop optional path parameters for this endpoint, making them
obligatory, therefore solving issues with base64 encoded input
containing slashes

Change-Id: I08172a105f94af7535d4122ccb685edf718c647d